Factor out control sys option tags to new file
